software/requirements.txt
numpy==1.26.4
pydantic==1.10.9
pygame==2.5.2
PyQt6==6.7.0
pyqtgraph==0.13.7
pyserial==3.5
scipy==1.13.1
